requirements.txt
# GeneChat2依赖包
# 完全符合论文实现的所有依赖

# ============ 核心依赖 ============
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0  # LoRA支持
accelerate>=0.20.0

# ============ DNA/生物信息学 ============
biopython>=1.81  # NCBI数据处理和序列操作

# ============ 数据处理 ============
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============ 评估指标 ============
nltk>=3.8  # BLEU和METEOR
sacrebleu>=2.3.0  # 替代BLEU实现
rouge-score>=0.1.2  # ROUGE指标（可选）

# ============ 训练和日志 ============
tensorboard>=2.13.0
tqdm>=4.65.0
wandb>=0.15.0  # 可选：实验跟踪

# ============ 配置和工具 ============
pyyaml>=6.0
omegaconf>=2.3.0  # 高级配置管理
dataclasses>=0.6; python_version < '3.7'

# ============ API和数据获取 ============
requests>=2.31.0
urllib3>=2.0.0

# ============ 可视化（可选） ============
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ============ 开发工具（可选） ============
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
ipython>=8.14.0
jupyter>=1.0.0

# ============ 性能优化 ============
flash-attn>=2.0.0; platform_system == "Linux"  # Flash Attention（Linux GPU）
deepspeed>=0.9.0  # 分布式训练（可选）
bitsandbytes>=0.41.0  # 量化支持（可选）

# ============ 其他工具 ============
scikit-learn>=1.3.0
h5py>=3.9.0
datasets>=2.13.0  # HuggingFace数据集
